chore: add extractFileID util
